move convert_observable_to_matrix() from utiltity to observable
